+++ interface.go
@@ -2,6 +2,7 @@ package gnfinder
 
 import (
 	"github.com/gnames/gnfinder/ent/output"
+	"github.com/gnames/gnlib/ent/gnvers"
 )
 
 type GNfinder interface {
@@ -9,5 +10,7 @@ type GNfinder interface {
 
 	GetConfig() Config
 
-	UpdateConfig(opts ...Option)
+	ChangeConfig(opts ...Option) GNfinder
+
+	GetVersion() gnvers.Version
 }